marche

+++ MementisLibrary/src/BSParameters.cpp
@@ -4,7 +4,7 @@
 
 
 BSParameters::BSParameters(int size, RateModel* r, PnlMat* matCorr, PnlVect * sigma, PnlVect* spot,
- PnlVect* mu, int hedgingDateNumber, std::map<int,int> mapDevises) {
+ PnlVect* mu, int hedgingDateNumber, std::map<int,double> mapDevises) {
 	this->sigma_ = sigma;
 	this->spot_ = spot;
 	this->size_ = size;
@@ -14,7 +14,6 @@ BSParameters::BSParameters(int size, RateModel* r, PnlMat* matCorr, PnlVect * si
 	this->Gn = pnl_vect_create(size_);
 	this->CorrelationMat = matCorr;
 	this->mapDevises = mapDevises;
-
 	// Factorisation de cholesky de la matrice de correlation
 	this->cholesky = pnl_mat_chol(CorrelationMat);
 }
