+++ MementisLibrary/src/BSParameters.hpp
@@ -21,9 +21,9 @@ public:
 	PnlVect *mu_;
 	PnlVect *Gn;
 	int hedgingDateNb_;
-	std::map<int,int> mapDevises;
+	std::map<int,double> mapDevises;
 	
-	BSParameters(int size, RateModel* r, PnlMat* matCorr, PnlVect * sigma, PnlVect* spot, PnlVect* mu, int hedgingDateNumber, std::map<int,int> mapDevises);
+	BSParameters(int size, RateModel* r, PnlMat* matCorr, PnlVect * sigma, PnlVect* spot, PnlVect* mu, int hedgingDateNumber, std::map<int,double> mapDevises);
 
 	~BSParameters();
 };
