+++ MementisLibrary/src/ProduitMementis.hpp
@@ -6,7 +6,7 @@ class ProduitMementis : public Produit
 public:
 	ProduitMementis();
 
-	ProduitMementis(FlowCalculator* flow, double maturity, int nbTimeSteps_, int nbAssets_ , int nbConstationDates, RateModel* interest, std::map<int,int> mapDevises);
+	ProduitMementis(FlowCalculator* flow, double maturity, int nbTimeSteps_, int nbAssets_ , int nbConstationDates, RateModel* interest, std::map<int,double> mapDevises);
 
 	virtual ~ProduitMementis();
 
@@ -14,7 +14,7 @@ public:
 	int nbAssets_;
 	int nbConstationDates;
 	double PayOffReel;
-	std::map<int,int> mapDevises;
+	std::map<int,double> mapDevises;
 
 	double payoff(const PnlMat *path);
 };
