+++ MementisLibrary/src/RateModel.hpp
@@ -4,7 +4,7 @@ class RateModel
 {
 public:
 
-	virtual double getRate(double t) = 0;
-	virtual double integrateRate(double t1, double t2, double indice = 3) =0;
+	virtual double getRate(int t) = 0;
+	virtual double integrateRate(double t1, double t2, int indice = 3) =0;
 
 };
