+++ MementisLibrary/src/Utils.hpp
@@ -7,5 +7,5 @@ namespace Utils {
   void matriceCorrelation(PnlMat* matCorr, PnlRng* rng, double nbActions);
   void mu(PnlVect *mu);
   void sigma(PnlVect *sigma);
-  void mapsDevises(std::map<int,int>& devises);
+  void mapsDevises(std::map<int,double>& devises);
 };
